+++ tx-backend/src/test/java/org/eclipse/tractusx/traceability/integration/qualitynotification/investigation/InvestigationControllerFilterIT.java
@@ -101,6 +101,112 @@ class InvestigationControllerFilterIT extends IntegrationTestSpecification {
     }
 
     @Test
+    void givenInvestigations_whenInvalidLocalDate_thenReturnBadRequest() throws JoseException {
+        // given
+        investigationNotificationSupport.defaultInvestigationsStored();
+        String filter = "?filter=createdDate,AT_LOCAL_DATE,2023-10-1111111,AND";
+
+        // when/then
+        given()
+                .header(oAuth2Support.jwtAuthorization(ADMIN))
+                .param("page", "0")
+                .param("size", "10")
+                .contentType(ContentType.JSON)
+                .when()
+                .get("/api/investigations" + filter)
+                .then()
+                .statusCode(400);
+    }
+
+    @Test
+    void givenInvestigations_whenTargetDateAtLocalDate_thenExpectedResult() throws JoseException {
+        // given
+        investigationNotificationSupport.defaultInvestigationsStored();
+        String filter = "?filter=targetDate,AT_LOCAL_DATE,2023-11-10,AND";
+
+        // when/then
+        given()
+                .header(oAuth2Support.jwtAuthorization(ADMIN))
+                .param("page", "0")
+                .param("size", "10")
+                .contentType(ContentType.JSON)
+                .when()
+                .get("/api/investigations" + filter)
+                .then()
+                .statusCode(200)
+                .body("page", Matchers.is(0))
+                .body("pageSize", Matchers.is(10))
+                .body("totalItems", Matchers.is(2))
+                .body("content", Matchers.hasSize(2));
+    }
+
+    @Test
+    void givenInvestigations_whenProvideFilterWithSeverityCritical_thenReturnAllCritical() throws JoseException {
+        // given
+        investigationNotificationSupport.defaultInvestigationsStored();
+
+        // when/then
+        given()
+                .header(oAuth2Support.jwtAuthorization(ADMIN))
+                .param("page", "0")
+                .param("size", "10")
+                .contentType(ContentType.JSON)
+                .when()
+                .param("filter", "severity,EQUAL,CRITICAL,AND")
+                .get("/api/investigations")
+                .then()
+                .statusCode(200)
+                .body("page", Matchers.is(0))
+                .body("pageSize", Matchers.is(10))
+                .body("totalItems", Matchers.is(2))
+                .body("content", Matchers.hasSize(2));
+    }
+
+    @Test
+    void givenInvestigations_whenProvideFilterCreatedBy_thenReturnAllCritical() throws JoseException {
+        // given
+        investigationNotificationSupport.defaultInvestigationsStored();
+
+        // when/then
+        given()
+                .header(oAuth2Support.jwtAuthorization(ADMIN))
+                .param("page", "0")
+                .param("size", "10")
+                .contentType(ContentType.JSON)
+                .when()
+                .param("filter", "createdBy,STARTS_WITH,BPNL00000001O,AND")
+                .get("/api/investigations")
+                .then()
+                .statusCode(200)
+                .body("page", Matchers.is(0))
+                .body("pageSize", Matchers.is(10))
+                .body("totalItems", Matchers.is(4))
+                .body("content", Matchers.hasSize(4));
+    }
+
+    @Test
+    void givenInvestigations_whenProvideFilterCreatedByName_thenReturnAllCritical() throws JoseException {
+        // given
+        investigationNotificationSupport.defaultInvestigationsStored();
+
+        // when/then
+        given()
+                .header(oAuth2Support.jwtAuthorization(ADMIN))
+                .param("page", "0")
+                .param("size", "10")
+                .contentType(ContentType.JSON)
+                .when()
+                .param("filter", "createdByName,STARTS_WITH,Car,AND")
+                .get("/api/investigations")
+                .then()
+                .statusCode(200)
+                .body("page", Matchers.is(0))
+                .body("pageSize", Matchers.is(10))
+                .body("totalItems", Matchers.is(4))
+                .body("content", Matchers.hasSize(4));
+    }
+
+    @Test
     void givenAlerts_whenProvideDateRangeFilters_thenReturnExpectedResult() throws JoseException {
         // given
         investigationNotificationSupport.defaultInvestigationsStored();