+++ CHANGELOG.md
@@ -13,6 +13,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Added missing translations
 - support for date ranges BEFORE_LOCAL_DATE and AFTER_LOCAL_DATE providing both will cause filter result to return only relevant date ranges
 - added supported searchCriteriaFieldsMappers for investigations, alerts, assetsAsBuilt and assetAsPlanned related endpoints
+- handling for invalid LocalDate string provided in filterCriteria for date related filtering
+- support for filtering join tables for Specification searchCriteria
 
 ### Changed
 - Updated user manual to reflect the table column settings feature